[QUIC]Change byte counting data type in watermark accounting to explicitly use uint64_t (#23115)
Commit Message: Change byte counting data type in watermark accounting to explicitly use uint64_t
Additional Description: when size_t is mapped to uint32_t, int64_t = uint32_t - uint32_t could potentially result in an unsigned wrap around, and causes error.
Risk Level: Low
Testing: unit test
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a
Fixes #23097

+++ source/common/quic/envoy_quic_stream.h
@@ -87,7 +87,7 @@ public:
   }
 
   // SendBufferMonitor
-  void updateBytesBuffered(size_t old_buffered_bytes, size_t new_buffered_bytes) override {
+  void updateBytesBuffered(uint64_t old_buffered_bytes, uint64_t new_buffered_bytes) override {
     if (new_buffered_bytes == old_buffered_bytes) {
       return;
     }
